Add Digimap CI script as internal script:
We have been adding the CI script as an external dependency but it
doesn't display the help. Installing from a GIT repo also seems to
add installation problems as the script doing the work is missing,
this is the big problem. To make life simpler, we also have more
pressing problems, I decided to copy the source to the scripts
directory and reference it as an internal script instead. This may
change if we figure out the other problems as it would be better for
it to be an external script.