Add interface2 application to whitelist:
While developing the new interface, we need to deploy it along-side
the existing interface. We have elected to run it under a different
context i.e. interface2. I have already updated the Jenkins jobs
involved and have now updated the script to permit the bot to work
with this app.