Prefix version:
Not sure why as others seem to use ndenv with no problem but the
hubot node version was not been set, even though this works in other
projects. I've changed it to be what is in the docs to see if this
makes any difference.

The problem is we don't all use ndenv, some use nenv which uses the
same file .node-version but the content is missing prefix.